// api/models/post-data-with-writer-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Social Media API
 * The Social Media API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TranslateResultDto } from './translate-result-dto';
import { WriterDataDto } from './writer-data-dto';
 /**
 * 
 *
 * @export
 * @interface PostDataWithWriterDto
 */
export interface PostDataWithWriterDto {

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    id: string;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    user: string;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    postType: PostDataWithWriterDtoPostTypeEnum;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    postStatus: PostDataWithWriterDtoPostStatusEnum;

    /**
     * @type {TranslateResultDto}
     * @memberof PostDataWithWriterDto
     */
    content: TranslateResultDto;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    blurHash: string;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    thumbnail?: string;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    width: number;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    height: number;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    ratio: number;

    /**
     * @type {Array<string>}
     * @memberof PostDataWithWriterDto
     */
    hashtags: Array<string>;

    /**
     * @type {string}
     * @memberof PostDataWithWriterDto
     */
    url: string;

    /**
     * @type {Date}
     * @memberof PostDataWithWriterDto
     */
    publishedAt: Date;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    likes: number;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    comments: number;

    /**
     * @type {number}
     * @memberof PostDataWithWriterDto
     */
    views: number;

    /**
     * @type {boolean}
     * @memberof PostDataWithWriterDto
     */
    liked: boolean;

    /**
     * @type {WriterDataDto}
     * @memberof PostDataWithWriterDto
     */
    writer: WriterDataDto;
}

/**
 * @export
 * @enum {string}
 */
export enum PostDataWithWriterDtoPostTypeEnum {
    Image = 'image',
    Video = 'video'
}
/**
 * @export
 * @enum {string}
 */
export enum PostDataWithWriterDtoPostStatusEnum {
    Inprogress = 'inprogress',
    Published = 'published',
    Failed = 'failed'
}